Support apps which are on Python2 but will have virtualenv on Python3
Some apps still on python2 so until they are migrated I introduced additional logic that allows us to have deploy_ve on Python3 even with compat4. Virtualenv lib upgrade to version 20+ cause major changes in our virtualenv.py, previous  logic related to relocataeble stuff  was purged so remove all of that, newer version doesn't need it

+++ yodeploy/virtualenv.py
@@ -1,16 +1,10 @@
-from __future__ import absolute_import
-
 import hashlib
 import logging
 import os
 import subprocess
 import sys
 import shutil
-import sysconfig
 import tarfile
-
-from pkg_resources import parse_version
-import virtualenv
 
 log = logging.getLogger(__name__)
 
@@ -22,19 +16,26 @@ def sha224sum(filename):
     return m.hexdigest()
 
 
-def get_python_version(compat):
-    """Return the version of Python that an app will use.
-
-    Based on the compat level.
-    """
-    if compat < 5:
+def get_python_version(compat=None, is_deploy=False):
+    python3 = shutil.which('python3')
+    if not python3:
+        log.error(
+            'python3 not found in PATH%s' % (
+                'for deploy VE' if is_deploy else ''
+            )
+        )
+        sys.exit(1)
+    python3_version = subprocess.check_output(
+        [
+            python3,
+            '-c', 'import sys; print(".".join(map(str, sys.version_info[:2])))'
+        ]
+    ).decode().strip()
+    if is_deploy:
+        return python3_version
+    if compat == 4:
         return '2.7'
-    if sys.version_info.major == 3:
-        return sysconfig.get_python_version()
-    return subprocess.check_output((
-        'python3', '-c',
-        'import sysconfig; print(sysconfig.get_python_version())'
-    )).strip()
+    return python3_version
 
 
 def get_id(filename, python_version, platform):
@@ -71,65 +72,88 @@ def create_ve(
         req_file='requirements.txt',
         verify_req_install=True):
     log.info('Building virtualenv')
-    ve_dir = os.path.join(app_dir, 'virtualenv')
+    ve_dir = os.path.abspath(os.path.join(app_dir, 'virtualenv'))
     req_file = os.path.join(os.path.abspath(app_dir), req_file)
+    # Ensure a clean slate by removing any existing VE directory
+    # to avoid conflicts or stale files from previous builds.
+    if os.path.exists(ve_dir):
+        shutil.rmtree(ve_dir)
 
-    # The venv module makes a lot of our reclocateability problems go away, so
-    # we only use the virtualenv library on Python 2.
-    if python_version.startswith('3.'):
-        subprocess.check_call((
-            'python%s' % python_version, '-m', 'venv', ve_dir))
-        pip_install(ve_dir, pypi, '-U', 'pip')
-    elif python_version == sysconfig.get_python_version():
-        virtualenv.create_environment(ve_dir, site_packages=False)
-    else:
-        subprocess.check_call((
-            sys.executable, virtualenv.__file__.rstrip('c'),
-            '-p', 'python%s' % python_version,
-            '--no-site-packages', ve_dir))
+    python_path = shutil.which('python%s' % python_version)
+    if not python_path:
+        log.error('Python %s not found in PATH', python_version)
+        sys.exit(1)
+    log.debug('Using Python path: %s', python_path)
 
-    pip_install(ve_dir, pypi, '-U', 'setuptools')
-    pip_install(ve_dir, pypi, 'wheel')
-
-    log.info('Installing requirements')
-    pip_install(ve_dir, pypi, '-r', req_file)
-
-    if verify_req_install:
-        log.info('Verifying requirements were met')
-        check_requirements(ve_dir)
-
-    relocateable_ve(ve_dir, python_version)
-
-    ve_id = get_id(os.path.join(app_dir, req_file), python_version, platform)
-    with open(os.path.join(ve_dir, '.hash'), 'w') as f:
-        f.write(ve_id)
-        f.write('\n')
-
-    log.info('Building virtualenv tarball')
-    cwd = os.getcwd()
-    os.chdir(app_dir)
-    t = tarfile.open('virtualenv.tar.gz', 'w:gz')
     try:
-        t.add('virtualenv')
-    finally:
-        t.close()
-        os.chdir(cwd)
+        if python_version.startswith('3.'):
+            log.debug('Creating Python 3 VE with venv: %s', ve_dir)
+            subprocess.check_call([python_path, '-m', 'venv', ve_dir])
+        else:
+            log.debug('Creating Python 2.7 VE with virtualenv: %s', ve_dir)
+            subprocess.check_call([
+                sys.executable, '-m', 'virtualenv', ve_dir,
+                '--python', python_path,
+                '--no-download'],
+                stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        python_bin = os.path.join(ve_dir, 'bin', 'python')
+        if not os.path.exists(python_bin):
+            log.error('VE created but bin/python missing at %s', python_bin)
+            log.debug('VE contents: %s', os.listdir(ve_dir))
+            raise FileNotFoundError(f"bin/python not found in {ve_dir}")
+        log.debug('VE created, contents: %s', os.listdir(ve_dir))
+
+        pip_install(ve_dir, pypi, '-U', 'pip')
+        pip_install(ve_dir, pypi, '-U', 'setuptools')
+        pip_install(ve_dir, pypi, 'wheel')
+        log.info('Installing requirements')
+        pip_install(ve_dir, pypi, '-r', req_file)
+        if verify_req_install:
+            log.info('Verifying requirements were met')
+            check_requirements(ve_dir)
+        relocateable_ve(ve_dir, python_version)
+        ve_id = get_id(
+            os.path.join(app_dir, req_file),
+            python_version, platform
+        )
+        with open(os.path.join(ve_dir, '.hash'), 'w') as f:
+            f.write(ve_id)
+            f.write('\n')
+        log.info('Building virtualenv tarball')
+        cwd = os.getcwd()
+        os.chdir(app_dir)
+        t = tarfile.open('virtualenv.tar.gz', 'w:gz')
+        try:
+            t.add('virtualenv')
+        finally:
+            t.close()
+            os.chdir(cwd)
+    except subprocess.CalledProcessError as e:
+        log.error('Failed to create VE: %s', e)
+        if e.stdout:
+            log.error('Stdout: %s', e.stdout.decode())
+        if e.stderr:
+            log.error('Stderr: %s', e.stderr.decode())
+        if os.path.exists(ve_dir):
+            shutil.rmtree(ve_dir)
+        sys.exit(1)
 
 
 def pip_install(ve_dir, pypi, *arguments):
     sub_log = logging.getLogger(__name__ + '.pip')
-
-    cmd = [os.path.join('bin', 'python'), '-m', 'pip', 'install']
+    cmd = [os.path.join(ve_dir, 'bin', 'python'), '-m', 'pip', 'install']
     if pypi:
         cmd += ['--index-url', pypi]
     cmd += arguments
-
-    p = subprocess.Popen(cmd, cwd=ve_dir, stdout=subprocess.PIPE)
-    for line in iter(p.stdout.readline, b''):
-        line = line.decode('utf-8').strip()
-        sub_log.info('%s', line)
-
-    if p.wait() != 0:
+    p = subprocess.Popen(
+        cmd, cwd=ve_dir, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+    )
+    stdout, stderr = p.communicate()
+    for line in stdout.decode().splitlines():
+        sub_log.info(line.strip())
+    if p.returncode != 0:
+        for line in stderr.decode().splitlines():
+            sub_log.error(line.strip())
         log.error('pip exited non-zero (%i)', p.returncode)
         sys.exit(1)
 
@@ -137,48 +161,30 @@ def pip_install(ve_dir, pypi, *arguments):
 def check_requirements(ve_dir):
     """Run pip check"""
     p = subprocess.Popen(
-        (os.path.join(ve_dir, 'bin', 'python'), '-m', 'pip', 'check'),
-        stdout=subprocess.PIPE)
-    out, _ = p.communicate()
-
+        [os.path.join(ve_dir, 'bin', 'python'), '-m', 'pip', 'check'],
+        stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    out, err = p.communicate()
     if p.returncode == 0:
-        log.info(out)
+        log.info(out.decode().strip())
     else:
-        log.error(out or 'Requirements check failed for unknown reasons')
+        log.error(err.decode().strip() or 'Requirements check failed')
         sys.exit(1)
 
 
 def relocateable_ve(ve_dir, python_version):
-    # We can only call into the virtualenv module for operating on a
-    # Python 2 virtualenv, in Python 2.
-    if sys.version_info.major == 3 and python_version == '2.7':
-        cmd = [
-            os.path.join(ve_dir, 'bin', 'python'), '-c',
-            'import sys; sys.path.append("%(ve_path)s"); '
-            'sys.path.append("%(my_path)s"); '
-            'from yodeploy import virtualenv; '
-            'virtualenv.relocateable_ve("%(ve_dir)s", "%(python_version)s")'
-            % {
-                've_path': os.path.dirname(virtualenv.__file__),
-                'my_path': os.path.join(os.path.dirname(__file__), '..'),
-                've_dir': ve_dir,
-                'python_version': python_version,
-            }]
-        subprocess.check_call(cmd)
-        return
-
     log.debug('Making virtualenv relocatable')
-
     # Python 3 venv virtualenvs don't have these problems
+    if python_version.startswith('3.'):
+        return
     if python_version == '2.7':
-        virtualenv.make_environment_relocatable(ve_dir)
         fix_local_symlinks(ve_dir)
-        remove_fragile_symlinks(ve_dir)
-
-    # Make activate relocatable, using approach taken in
-    # https://github.com/pypa/virtualenv/pull/236
+        remove_fragile_symlinks(ve_dir, python_version)
+    activate_path = os.path.join(ve_dir, 'bin', 'activate')
+    if not os.path.exists(activate_path):
+        log.error('activate script missing in %s', ve_dir)
+        return
     activate = []
-    with open(os.path.join(ve_dir, 'bin', 'activate')) as f:
+    with open(activate_path) as f:
         for line in f:
             line = line.strip()
             if line == 'deactivate () {':
@@ -187,18 +193,13 @@ def relocateable_ve(ve_dir, python_version):
                 activate += [
                     '# attempt to determine VIRTUAL_ENV in relocatable way',
                     'if [ ! -z "${BASH_SOURCE:-}" ]; then',
-                    '    # bash',
                     '    ACTIVATE_PATH="${BASH_SOURCE}"',
                     'elif [ ! -z "${DASH_SOURCE:-}" ]; then',
-                    '    # dash',
                     '    ACTIVATE_PATH="${DASH_SOURCE}"',
                     'elif [ ! -z "${ZSH_VERSION:-}" ]; then',
-                    '    # zsh',
                     '    ACTIVATE_PATH="$0"',
                     'elif [ ! -z "${KSH_VERSION:-}" ] '
                          '|| [ ! -z "${.sh.version:}" ]; then',
-                    '    # ksh - we have to use history, '
-                                'and unescape spaces before quoting',
                     '    ACTIVATE_PATH="$(history -r -l -n | head -1 | sed -e '
                         '\'s/^[\\t ]*\\(\\.\\|source\\) *//;s/\\\\ / /g\')"',
                     'elif [ "$(basename "$ACTIVATE_PATH_FALLBACK")" == '
@@ -220,7 +221,7 @@ def relocateable_ve(ve_dir, python_version):
                     'unset ACTIVATE_PATH_FALLBACK',
                 ]
             activate.append(line)
-    with open(os.path.join(ve_dir, 'bin', 'activate'), 'w') as f:
+    with open(activate_path, 'w') as f:
         f.write('\n'.join(activate))
 
 
@@ -240,7 +241,7 @@ def fix_local_symlinks(ve_dir):
         os.symlink(target, symlink)
 
 
-def remove_fragile_symlinks(ve_dir):
+def remove_fragile_symlinks(ve_dir, python_version):
     """Remove symlinks to parent virtualenv.
 
     When we create a virtualenv with a Python from another virtualenv, we don't
@@ -250,15 +251,16 @@ def remove_fragile_symlinks(ve_dir):
     if getattr(sys, 'real_prefix', sys.prefix) == sys.prefix:
         return
     ve_prefix = sys.prefix
-    home_dir, lib_dir, inc_dir, bin_dir = virtualenv.path_locations(ve_dir)
+    home_dir = os.path.abspath(ve_dir)
+    lib_dir = os.path.join(home_dir, 'lib', 'python' + python_version)
+    if not os.path.exists(lib_dir):
+        return
     for fn in os.listdir(lib_dir):
         path = os.path.join(lib_dir, fn)
         if not os.path.islink(path):
             continue
-
         dest = os.readlink(path)
         if not dest.startswith(ve_prefix):
             continue
-
         log.debug('Removing fragile symlink %s -> %s', path, dest)
         os.unlink(path)
